requirements.txt
assemblyai
moviepy
reportlab
python-dotenv